react work: handle null summary, basic styling

+++ Server.AssetsFlow/Journal/JTA.Journal.Seeder/DatabaseSeeder.cs
@@ -17,7 +17,7 @@ namespace HsR.Journal.Seeder
         internal async Task SeedAsync()
         {
             // Check if any data exists in a specific table to avoid reseeding
-            //if (!await _dbContext.TradeComposites.AnyAsync())
+            if (!await _dbContext.TradeComposites.AnyAsync())
             {
                 // Drop the database if it exists
                 dbContext.Database.EnsureDeleted();
